refactor version tracking
