+++ openedx_plugin/apps.py
@@ -12,7 +12,7 @@ from django.apps import AppConfig
 from edx_django_utils.plugins import PluginSettings, PluginURLs
 from openedx.core.djangoapps.plugins.constants import ProjectType, SettingsType
 
-from ..version import __version__
+from .version import __version__
 
 log = logging.getLogger(__name__)
 log.info("openedx_plugin %s", __version__)
