+++ setup.py
@@ -3,7 +3,6 @@ from __future__ import absolute_import, print_function, unicode_literals
 
 import os
 from setuptools import find_packages, setup
-from version import __version__
 
 # allow setup.py to be run from any path
 os.chdir(os.path.normpath(os.path.join(os.path.abspath(__file__), os.pardir)))
@@ -48,7 +47,7 @@ print("requirements found: {requirements}".format(requirements=load_requirements
 
 setup(
     name="example-plugin",
-    version=__version__,
+    version="0.0.2",
     packages=find_packages(),
     package_data={"": ["*.html"]},  # include any Mako templates found in this repo.
     include_package_data=True,
